Crate rune_tracing

Minimum support: Rust 1.81+.

Visit the site ğŸŒ — Read the book ğŸ“–

Rune tracing shims for the Rune Language, an embeddable dynamic programming language for Rust.

Modules

event

Events represent single points in time during the execution of a program.

field

Span and Event key-value data.

instrument

Attach a span to a std::future::Future.

span

Spans represent periods of time in which a program was executing in a particular context.

Macros

debug

Constructs an event at the debug level.

enabled

Checks whether a span or event is enabled based on the provided metadata.

error

Constructs an event at the error level.

event

Constructs a new Event.

info

Constructs an event at the info level.

span

Constructs a new span.

trace

Constructs an event at the trace level.

warn

Constructs an event at the warn level.

Structs

Level

Describes the level of verbosity of a span or event.

Attribute Macros

instrument

Instruments a function to create and enter a tracing span every time the function is called.

instrument_ast